+++ lib/src/models/view_configurations/multi_day_configurations/multi_day_view_configuration.dart
@@ -6,6 +6,9 @@ abstract class MultiDayViewConfiguration extends ViewConfiguration {
     int? numberOfDays,
     int? firstDayOfWeek,
     bool? paintWeekNumber,
+    double? initialHeightPerMinute,
+    bool showHeader = true,
+    double hourLineLeftMargin = 56,
     required double timelineWidth,
     required double daySeparatorLeftOffset,
     required Duration horizontalStepDuration,
@@ -45,6 +48,11 @@ abstract class MultiDayViewConfiguration extends ViewConfiguration {
 
     _enableRescheduling = enableRescheduling;
     _enableResizing = enableResizing;
+
+    _showHeader = showHeader;
+    _hourLineLeftMargin = hourLineLeftMargin;
+
+    _initialHeightPerMinute = initialHeightPerMinute ?? 0.7;
 
     assert(
       startHour >= 0 && startHour <= 23,
@@ -89,6 +97,13 @@ abstract class MultiDayViewConfiguration extends ViewConfiguration {
   double get daySeparatorLeftOffset => _daySeparatorLeftOffset;
   set daySeparatorLeftOffset(double value) {
     _daySeparatorLeftOffset = value;
+    notifyListeners();
+  }
+
+  late double _hourLineLeftMargin;
+  double get hourLineLeftMargin => _hourLineLeftMargin;
+  set hourLineLeftMargin(double value) {
+    _hourLineLeftMargin = value;
     notifyListeners();
   }
 
@@ -226,6 +241,16 @@ abstract class MultiDayViewConfiguration extends ViewConfiguration {
     _endHour = value;
     notifyListeners();
   }
+
+  late bool _showHeader;
+  bool get showHeader => _showHeader;
+  set showHeader(bool value) {
+    _showHeader = value;
+    notifyListeners();
+  }
+
+  late double _initialHeightPerMinute;
+  double get heightPerMinute => _initialHeightPerMinute;
 
   bool get customStartEndHour => _startHour != 0 || _endHour != 24;
 }