+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+* Added initialHeightPerMinute.
+* Fixed MultiDayView header shadow issue.
+* Added hourLineLeftMargin to control the left margin of the hour line.
+
+## 0.1.9
+* Fixed start and end hour issue.
+
 ## 0.1.8
 * Fixed enable resize option not working.